refactor(api-client): Migrate Node.js demo to TypeScript
