+++ packages/api-client/package.json
@@ -30,6 +30,7 @@
     "browser-sync": "2.26.7",
     "concurrently": "5.0.0",
     "cross-env": "6.0.3",
+    "dotenv": "8.2.0",
     "istanbul": "1.1.0-alpha.1",
     "jasmine": "3.5.0",
     "karma": "4.4.1",
@@ -39,13 +40,13 @@
     "karma-sourcemap-loader": "0.3.7",
     "karma-spec-reporter": "0.0.32",
     "nock": "11.7.0",
-    "optimist": "0.6.1",
     "react": "16.12.0",
     "react-dom": "16.12.0",
     "react-hot-loader": "4.12.18",
     "rimraf": "3.0.0",
     "sinon": "7.5.0",
     "sinon-har-server": "0.3.0",
+    "ts-node": "8.5.2",
     "typescript": "3.6.4",
     "webpack": "4.41.2",
     "webpack-cli": "3.3.10",
@@ -69,7 +70,7 @@
     "coverage": "istanbul cover --report html ./node_modules/jasmine/bin/jasmine.js",
     "dist": "yarn clean && yarn build",
     "start": "yarn build:node && concurrently \"tsc -w\" \"webpack -w\" \"browser-sync start -c bs-config.js\"",
-    "start:node": "yarn build:node && node ./index.js",
+    "start:node": "ts-node index.ts",
     "test": "yarn test:node && yarn test:browser",
     "test:project": "yarn dist && yarn test",
     "test:browser": "webpack && karma start",
